Rails 3.1.3 support
The implementation of ActiveRecord::Base.arel_attributes_values changed
between 3.1.3 and 3.2 -- they stopped storing the serializer in
.serialized_attributes and instead store it directly in the @attributes
array. This commit changes the override for arel_attributes_values to
reflect the old way of encoding serialized values for storage in the
database.

There are also now more-specific versions of dependencies referenced in
the .gemspec to ensure binding to ActiveRecord 3.1.3. combustion has a
conflicting dependency on "rails >= 3.0.0" so this commit also adds a
specific development dependency for 3.1.3.